moved Newsletter to a form
